add dtype for converting classes to tensor

+++ ultralytics/utils/ops.py
@@ -245,8 +245,7 @@ def non_max_suppression(
     if isinstance(prediction, (list, tuple)):  # YOLOv8 model in validation model, output = (inference_out, loss_out)
         prediction = prediction[0]  # select only inference output
     if classes is not None:
-        classes = torch.tensor(classes, device=prediction.device)
-
+        classes = torch.tensor(classes, device=prediction.device, dtype=torch.long)
     if prediction.shape[-1] == 6 or end2end:  # end-to-end model (BNC, i.e. 1,300,6)
         output = [pred[pred[:, 4] > conf_thres][:max_det] for pred in prediction]
         if classes is not None:
